+++ example/lib/examples/search_page.dart
@@ -1,48 +1,27 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:yandex_mapkit/yandex_mapkit.dart';
 import 'package:yandex_mapkit_example/examples/widgets/control_button.dart';
 import 'package:yandex_mapkit_example/examples/widgets/map_page.dart';
 
-class SearchPage extends MapPage {
-  const SearchPage() : super('Search example');
+class SuggestionsPage extends MapPage {
+  const SuggestionsPage() : super('Suggestions example');
 
   @override
   Widget build(BuildContext context) {
-    return _SearchExample();
+    return _SuggestionsExample();
   }
 }
 
-class _SearchExample extends StatefulWidget {
+class _SuggestionsExample extends StatefulWidget {
   @override
-  _SearchExampleState createState() => _SearchExampleState();
+  _SuggestionsExampleState createState() => _SuggestionsExampleState();
 }
 
-class _SearchExampleState extends State<_SearchExample> {
-
+class _SuggestionsExampleState extends State<_SuggestionsExample> {
   TextEditingController queryController = TextEditingController();
-
-  final List<SearchResponse> responseByPages = [];
-
-  final Map<int,SearchSession> _sessions = {};
-
-  @override
-  void dispose() async {
-
-    super.dispose();
-
-    try {
-      for (var s in _sessions.values) {
-        await s.close();
-      }
-    } on SearchSessionException catch (e) {
-      print('Error: ${e.message}');
-    }
-
-    _sessions.clear();
-  }
+  String response = '';
 
   @override
   Widget build(BuildContext context) {
@@ -66,7 +45,7 @@ class _SearchExampleState extends State<_SearchExample> {
                     ),
                     ControlButton(
                       onPressed: () {
-                        search(queryController.text);
+                        querySuggestions(queryController.text);
                       },
                       title: 'Query'
                     ),
@@ -78,13 +57,7 @@ class _SearchExampleState extends State<_SearchExample> {
                   mainAxisAlignment: MainAxisAlignment.start,
                   children: <Widget>[
                     Flexible(
-                      child: Padding(
-                        padding: EdgeInsets.only(top: 20),
-                        child: Column(
-                          crossAxisAlignment: CrossAxisAlignment.start,
-                          children: _getList(),
-                        )
-                      ),
+                      child: Text(response)
                     ),
                   ],
                 ),
@@ -96,91 +69,19 @@ class _SearchExampleState extends State<_SearchExample> {
     );
   }
 
-  List<Widget> _getList() {
-
-    var list = <Widget>[];
-
-    for (var r in responseByPages) {
-
-      list.add(Text('Page: ${r.page}'));
-      list.add(Container(height: 20));
-
-      r.items.asMap().forEach((i, item) {
-        list.add(Text('Item $i: ${item.toponymMetadata!.formattedAddress}'));
-      });
-
-      list.add(Container(height: 20));
-    }
-
-    return list;
-  }
-
-  void search(String query) async {
-
-    print('Search query: $query');
-
-    responseByPages.clear();
-
-    var sessionWithResponse = await YandexSearch.searchByText(
-      searchText: query,
-      geometry: Geometry.fromBoundingBox(
-        BoundingBox(
-          southWest: Point(latitude: 55.76996383933034, longitude: 37.57483142322235),
-          northEast: Point(latitude: 55.785322774728414, longitude: 37.590924677311705),
-        )
-      ),
-      searchOptions: SearchOptions(
-        searchType: SearchType.geo,
-        geometry: false,
-      ),
-    );
-
-    var session = sessionWithResponse.session;
-
-    _sessions[session.id] = session;
-
-    var responseOrError = await sessionWithResponse.responseOrError;
-
-    await _handleResponse(session, responseOrError);
-
-    print('No more results available, closing session...');
-    await _closeSession(session);
-  }
-
-  Future<void> _closeSession(SearchSession session) async {
-
-    try {
-      await session.close();
-    } on SearchSessionException catch (e) {
-      print('Error: ${e.message}');
-    }
-
-    _sessions.remove(session.id);
-  }
-
-  Future<void> _handleResponse(SearchSession session, SearchResponseOrError responseOrError) async {
-
-    if (responseOrError.error != null) {
-      print('Error: ${responseOrError.error}');
-      await _closeSession(session);
-      return;
-    }
-
-    var response = responseOrError.response!;
-
-    print('Page ${response.page}: ${response.toString()}');
-
-    setState(() {
-      responseByPages.add(response);
-    });
-
-    try {
-      if (await session.hasNextPage()) {
-        print('Got ${response.found} items, fetching next page...');
-        await _handleResponse(session, await session.fetchNextPage());
+  Future<void> querySuggestions(String query) async {
+    final cancelListening = await YandexSuggest.getSuggestions(
+      address: query,
+      southWestPoint: const Point(latitude: 55.5143, longitude: 37.24841),
+      northEastPoint: const Point(latitude: 56.0421, longitude: 38.0284),
+      suggestType: SuggestType.geo,
+      suggestWords: true,
+      onSuggest: (List<SuggestItem> suggestItems) {
+        setState(() {
+          response = suggestItems.map((SuggestItem item) => item.title).join('\n');
+        });
       }
-    } on SearchSessionException catch (e) {
-      print('Error: ${e.message}');
-    }
+    );
+    await Future<dynamic>.delayed(const Duration(seconds: 3), () => cancelListening());
   }
 }
