+++ example/lib/main.dart
@@ -11,7 +11,7 @@ import 'package:yandex_mapkit_example/examples/polyline_page.dart';
 import 'package:yandex_mapkit_example/examples/polygon_page.dart';
 import 'package:yandex_mapkit_example/examples/rotation_page.dart';
 import 'package:yandex_mapkit_example/examples/search_page.dart';
-import 'package:yandex_mapkit_example/examples/suggestions_page.dart';
+import 'package:yandex_mapkit_example/examples/search_page2.dart';
 
 void main() {
   runApp(MaterialApp(home: MainPage()));
